changed all to floats and organized code

+++ classes/hittable.cuh
@@ -7,7 +7,7 @@ struct hit_record
 {
     point3 p; // 3D point of hit
     vec3 normal; // normal
-    double t; // t value in ray equation
+    float t; // t value in ray equation
     bool front_face;
 
     __device__ inline void set_face_normal(const ray& r, const vec3& outward_normal) {
@@ -19,7 +19,7 @@ struct hit_record
 class hittable
 {
 public:
-    __device__ virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const = 0;
+    __device__ virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
 };
 
 #endif
