+++ classes/ray.cuh
@@ -6,14 +6,12 @@
 class ray {
     public:
         __device__ ray() {}
-        __device__ ray(const point3& origin, const vec3& direction)
-            : orig(origin), dir(direction)
-        {}
+        __device__ ray(const point3& origin, const vec3& direction) : orig(origin), dir(direction) {}
 
         __device__ point3 origin() const  { return orig; }
         __device__ vec3 direction() const { return dir; }
 
-        __device__ point3 at(double t) const {
+        __device__ point3 at(float t) const {
             return orig + t*dir;
         }
